requirements.txt
aniso8601==8.0.0
bcrypt==3.1.6
blinker==1.4
certifi==2019.11.28
cffi==1.12.2
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==5.4.1
pycparser==2.19
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.25.8
virtualenv==16.4.3
waitress==1.4.3
Werkzeug==0.14.1
WTForms==2.2.1
